Add files via upload
Why do quaternions have to be so complicated. I spent like 5 hours trying to comprehend whatever the fuck they were and I learned a bunch of shit. For example that my code sucks and the angle that the quaternion moves by is inaccurate but that is a problem for future me :) .